调整系统权限相关接口

+++ src/main/java/com/simbest/boot/sys/repository/SysDictRepository.java
@@ -7,14 +7,14 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SysDictRepository extends BaseRepository<SysDict, Long> {
+public interface SysDictRepository extends BaseRepository<SysDict, Integer> {
 
     //    String SQL = "SELECT * FROM sys_dict WHERE parent_id=:id";
 //    @Query(value = SQL, nativeQuery = true)
 //    List<SysDict> findByParentId(@Param("id")Long parentId);
-    List<SysDict> findByParentId(Long parentId);
+    List<SysDict> findByParentId(Integer parentId);
 
-    List<SysDict> findByParentIdAndEnabledAndRemoved(Long parentId, Boolean enabled, Boolean removed);
+    List<SysDict> findByParentIdAndEnabledAndRemoved(Integer parentId, Boolean enabled, Boolean removed);
 
     List<SysDict> findByEnabled(Boolean enabled);
